+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/docker_kurtosis_backend_module_functions.go
@@ -102,7 +102,9 @@ func (backend *DockerKurtosisBackend) CreateModule(
 	shouldReleaseIp := true
 	defer func() {
 		if shouldReleaseIp {
-			freeIpAddrProvider.ReleaseIpAddr(ipAddr)
+			if err = freeIpAddrProvider.ReleaseIpAddr(ipAddr); err != nil {
+				logrus.Errorf("Error releasing IP address '%v'", ipAddr)
+			}
 		}
 	}()