+++ core/server/commons/enclave_data_directory/directory_existence_ensurer.go
@@ -11,10 +11,6 @@ import (
 )
 
 const (
-	// The system umask is a set of bits that are _subtracted_ from the perms when we create a file
-	// We really do want a 0777 directory (see comment below), so we have to set this to 0
-	umaskForCreatingDirectory = 0
-
 	// NOTE: It's very important that all directories created inside the enclave data directory are created with 0777
 	//  permissions, because:
 	//  a) the engine data directory (and containing enclave data dirs) are bind-mounted on the Docker host machine