+++ core/server/commons/enclave_data_directory/files_artifact_store_test.go
@@ -103,6 +103,7 @@ func TestFileStore_RemoveFileRemovesFileFromDisk(t *testing.T) {
 func TestFileStore_RemoveFileFailsForNonExistentUuid(t *testing.T) {
 	fileStore := getTestFileStore(t)
 	nonExistentUuid, err := newFilesArtifactUUID()
+	require.Nil(t, err)
 
 	err = fileStore.RemoveFile(nonExistentUuid)
 	require.NotNil(t, err)